feat: Cast Sum variables as decimal types for accurate Score calculation

+++ StudentGrades/Program.cs
@@ -38,10 +38,10 @@ namespace StudentGrades
             int zahirahSum = zahirah1 + zahirah2 + zahirah3 + zahirah4 + zahirah5;
             int jeongSum = jeong1 + jeong2 + jeong3 + jeong4 + jeong5;
 
-            decimal sophiaScore = sophiaSum / currentAssignments;
-            decimal nicolasScore = nicolasSum / currentAssignments;
-            decimal zahirahScore = zahirahSum / currentAssignments;
-            decimal jeongScore = jeongSum / currentAssignments;
+            decimal sophiaScore = (decimal) sophiaSum / currentAssignments;
+            decimal nicolasScore = (decimal) nicolasSum / currentAssignments;
+            decimal zahirahScore = (decimal) zahirahSum / currentAssignments;
+            decimal jeongScore = (decimal) jeongSum / currentAssignments;
 
             Console.WriteLine("Sophia: " + sophiaScore);
             Console.WriteLine("Nicolas: " + nicolasScore);
